requirements.txt
requests
flask
python-dotenv
gunicorn
flask-cors